requirements.txt
streamlit 
pymupdf 
python-dotenv
apify-client
google-generativeai
chromadb
sentence-transformers
pandas
numpy
plotly
langchain
langchain-community
langchain-google-genai